requirements.txt
Pillow==9.5.0
emoji
instabot
python-dotenv